[TECH] rename store entry point
