+++ src/front/app/main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { Router } from 'react-router';
-import appStore from './store/store';
+import appStore from './store';
 import routes from './routes';
 
 export default () => {
